Refactor tests: move helpers to utils.py + minor log change
